chore: add error handling for duplicate nodes or edges
