+++ srcs/parse/parse_reading.c
@@ -6,7 +6,7 @@
 /*   By: sucho <sucho@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/07 20:44:25 by sucho             #+#    #+#             */
-/*   Updated: 2023/08/13 22:56:25 by sucho            ###   ########.fr       */
+/*   Updated: 2023/08/13 23:14:49 by sucho            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -155,16 +155,4 @@ int	parse_check_nodeline(t_list **line_head, t_parse **parse)
 		// line_count++;
 	}
 	return (1);
-}
-
-void	parse_check_antnum(t_list **line_head, t_parse **parse)
-{
-	if (!(0 < ft_atoi((*line_head)->content) && \
-			ft_atoi((*line_head)->content) <= INT_MAX))
-	{
-		ft_putstr_fd("Error in first line: ant number is wrong \n", STDOUT_FILENO);
-		exit(1);
-	}
-	(*parse)->num_ants = ft_atoi((*line_head)->content);
-	(*line_head) = (*line_head)->next;
 }
